+++ thpsbot/helpers/autocomplete_helper.py
@@ -4,7 +4,8 @@ from discord import Interaction, app_commands
 
 
 async def get_pfp_filenames(
-    interaction: Interaction, current: str
+    interaction: Interaction,
+    current: str,
 ) -> list[app_commands.Choice[str]]:
     pfps = os.listdir("pfps/")
     matches = [f for f in pfps if current.lower() in f.lower()]